fixing Board composant name
